+++ step_run_script.go
@@ -14,6 +14,9 @@ import (
 	"go.opencensus.io/trace"
 )
 
+var MaxLogLengthExceeded = errors.New("maximum log length exceeded")
+var LogWriterTimeout = errors.New("log writer timeout")
+
 type runScriptReturn struct {
 	result *backend.RunResult
 	err    error
@@ -62,16 +65,20 @@ func (s *stepRunScript) Run(state multistep.StateBag) multistep.StepAction {
 		// implementation returns with the error too quickly for the ctx.Done()
 		// case branch below to catch it.
 		if errors.Cause(r.err) == gocontext.DeadlineExceeded {
+			state.Put("err", r.err)
 			logger.Info("hard timeout exceeded, terminating")
 			s.writeLogAndFinishWithState(preTimeoutCtx, ctx, logWriter, buildJob, FinishStateErrored, "\n\nThe job exceeded the maximum time limit for jobs, and has been terminated.\n\n")
 			return multistep.ActionHalt
 		}
 		if logWriter.MaxLengthReached() {
+			state.Put("err", MaxLogLengthExceeded)
 			s.writeLogAndFinishWithState(preTimeoutCtx, ctx, logWriter, buildJob, FinishStateErrored, "\n\nThe job exceeded the maximum log length, and has been terminated.\n\n")
 			return multistep.ActionHalt
 		}
 
 		if r.err != nil {
+			state.Put("err", r.err)
+
 			if !r.result.Completed {
 				logger.WithFields(logrus.Fields{
 					"err":       r.err,
@@ -98,12 +105,15 @@ func (s *stepRunScript) Run(state multistep.StateBag) multistep.StepAction {
 
 		return multistep.ActionContinue
 	case <-ctx.Done():
+		state.Put("err", ctx.Err())
+
 		if ctx.Err() == gocontext.DeadlineExceeded {
 			logger.Info("hard timeout exceeded, terminating")
 			s.writeLogAndFinishWithState(preTimeoutCtx, ctx, logWriter, buildJob, FinishStateErrored, "\n\nThe job exceeded the maximum time limit for jobs, and has been terminated.\n\n")
 			return multistep.ActionHalt
 		}
 		if logWriter.MaxLengthReached() {
+			state.Put("err", MaxLogLengthExceeded)
 			s.writeLogAndFinishWithState(preTimeoutCtx, ctx, logWriter, buildJob, FinishStateErrored, "\n\nThe job exceeded the maximum log length, and has been terminated.\n\n")
 			return multistep.ActionHalt
 		}
@@ -111,10 +121,12 @@ func (s *stepRunScript) Run(state multistep.StateBag) multistep.StepAction {
 		logger.Info("context was cancelled, stopping job")
 		return multistep.ActionHalt
 	case <-cancelChan:
+		state.Put("err", JobCancelledError)
 		s.writeLogAndFinishWithState(preTimeoutCtx, ctx, logWriter, buildJob, FinishStateCancelled, "\n\nDone: Job Cancelled\n\n")
 
 		return multistep.ActionHalt
 	case <-logWriter.Timeout():
+		state.Put("err", LogWriterTimeout)
 		s.writeLogAndFinishWithState(preTimeoutCtx, ctx, logWriter, buildJob, FinishStateErrored, fmt.Sprintf("\n\nNo output has been received in the last %v, this potentially indicates a stalled build or something wrong with the build itself.\nCheck the details on how to adjust your build configuration on: https://docs.travis-ci.com/user/common-build-problems/#Build-times-out-because-no-output-was-received\n\nThe build has been terminated\n\n", s.logTimeout))
 
 		if s.skipShutdownOnLogTimeout {